Rework the config option to an enum
Summary:
After discussing our needs for easy mypy migration with Maggie and Zeina, I concluded
that the initially planned behavior of only allowing two modes (check and infer return
type vs skip and infer any) won't be sufficient: Zeina says that *half* of the mypy
primer projects use mypy's `check_untyped_defs` flag, which causes MyPy to infer Any
but still check the function body.

While I very much hope that we will be able to discourage this behavior, the extremely
high prevalence of both the legacy PEP 484 behavior of skipping check *and* the
`check_untyped_defs` behavior is enough to convince me that if pain-free migration is
a real priority, we must be able to emulate both mypy modes as well as our normal
Pyrefly-like mode.

This diff prepares by making that change in the config (it remains a no-op in terms of
actual type check as of this diff, we still are not respecting the option yet).

----

Note: I think it's still open to debate whether we really want three options here, but
I wanted to go ahead and make the change before implementing it because *if* we need
three options, I need the implementation to be done in a way that doesn't couple the
type check vs return type inference behaviors too tightly.

It's pretty easy to flip the behavior back to a two-way option once it's implemented if
we decide we don't actually need this.

Reviewed By: yangdanny97

Differential Revision: D74441070

fbshipit-source-id: a3114a9250c1672230bb0702c42cf5b4ff0ffc09