Better return type inference and checking for "stub" functions
Summary:
I noticed that we were using a bit of a hack (infer a return type of Never) to make sure we never report bad return type errors in functions with a `...` body, functions decorated with `abstractmethod`, or functions in `.pyi` files. This diff tries to handle these cases in a more principled way:

* All functions in `.pyi` files get an inferred return type of `Any`.
* Functions with a `...` body or an `abstractmethod` decoration get their return type inferred normally, but we use the return type annotation, if one exists, without checking it against the inferred type.

An important effect of this change is that a function with a `...` body and no return type annotation gets a return type of `Any` in a `.pyi` file and `None` in a `.py` file, accurately reflecting runtime behavior in the latter case.

Reviewed By: stroxler

Differential Revision: D74432310

fbshipit-source-id: 1e37ab554538bf8b73e33af771cf9aa31182ab5b