update border color for light mode + add corner radius
Summary:
the border color was previously hidden in light mode since it camouflaged with the background. This diff updates the background for light mode so that it's actually visible.
The UI also looked a little chunky since the inner box was square w/o corner radius and the border had corner radius. We add corner radius to inner component for consistency. This also gets rid of the blue border that shows up

Reviewed By: stroxler

Differential Revision: D74518001

fbshipit-source-id: 966b6fe170190b6accda96e074fc0bc7ba7492b4