unify handling of Unpack
Summary:
Previously, we had an Unpack qualifier (used for *args and **kwargs annotations) and a Type::Unpack (used everywhere else).

There was some weird logic to check and wrap the former into the latter, but if we handle everything as Type::Unpack we can actually delete the qualifier entirely and everything still works.

fixes https://github.com/facebook/pyrefly/issues/180

Reviewed By: rchen152

Differential Revision: D74520929

fbshipit-source-id: a959b58cbfba5e7d90bada553416fb2aba11a610