Add a test case of self attrs in untyped methods
Summary:
We want to make sure that when we implement skip-and-infer-return-any
mode for untyped defs, we remain aware of the existence of attributes
defined in such methods (anything else might produce an overly large
number of false positives on largely-untyped codebases).

Let's add a test case for this ahead of time.

Reviewed By: ndmitchell

Differential Revision: D74537914

fbshipit-source-id: 898337ae1cf6866fba3ad03b75098f311f2edc49