Pass around search path as an iterator
Summary:
In the following diffs, I'm going to split up the search path so that parts with different sources (e.g., command line vs config file) are stored separately. So that we don't have to repeatedly .collect() a combined search path repeatedly, make various functions on search paths take an Iterator.

My goal in doing this is to make import errors more informative; see the summary of D74554698 for an example of the final state I'm working towards.

Reviewed By: stroxler

Differential Revision: D74540850

fbshipit-source-id: 6d9e5bfaccfabe9561ea3c5114fa3e373211bb50