Track source of site_package_path more precisely
Summary:
Tracks where a config's site_package_path came from (command line, config file, or interpreter). For now, I've added this info to dump-config (mainly so I could manually test it), but the eventual goal is to incorporate this into the [import-error] error message.

See the summary of D74554698 for an example of the final state I'm working toward.

Reviewed By: stroxler

Differential Revision: D74539162

fbshipit-source-id: c94d3095a5f0638a6059f439a1754ed7334bac54