Rename ConfigFile::search_path to search_path_from_file
Summary:
Once I split up the search path, this field will contain only the portion of it that is parsed from the config file; preemptively rename it.

My goal in doing this is to make import errors more informative; see the summary of D74554698 for an example of the final state I'm working towards.

Reviewed By: stroxler

Differential Revision: D74540854

fbshipit-source-id: a4bb5e428602608391c236c5ed9747c54a2bf4fe