fix dark mode issue
Summary:
## Problem
I tried to add logic into the svg file to render differently depending on if we are in dark mode or not. This doesn't seem to work well on the actual pyrefly.org despite working locally.

## Solution
This change modifes this so that the branching for dark mode is a check in `custom.css` instead, and we have one svg for both light and dark mode for each symbol

Reviewed By: grievejia

Differential Revision: D74598930

fbshipit-source-id: 2e2483b7ea3244f8ec3c731505e26bb9f7591852