Update landing page header to use ThemedImage
Summary:
## Problem
`useColorMode` is returning a stale color during the initial load of the website. This problem didn't pop up in local development and only seem to be a problem in the prod website (pyrefly.org). This leads to the pyrefly logo to use the dark mode variant even in light mode.
 {F1977884472}

This seem to be a known issue https://github.com/facebook/docusaurus/issues/7986, recommended fix is it to use `ThemedImage` if possible.

## Solution
This diff updates the logo used in landing page header to use `ThemedHeader`

Reviewed By: grievejia

Differential Revision: D74598188

fbshipit-source-id: b7a113a20ae256ab507de55273ddd2976624ae2b