Refactor: create `Decorators` data structure
Summary:
I need a good way to track whether the `no_type_check` decorator is
present, which is not a `Key` level computation, in order to implement
the logic to skip type checking decorated functions.

But to do this without making a mess of the type function code, I really
need a single entity representing the result of decorator analysis, which
will need to be a bunch of `Idx<Key>` values *plus* a boolean indicating
whether we saw the `no_type_check` decorator.

This diff preps for that by pulling the existing code into a data structure.

Reviewed By: grievejia

Differential Revision: D74582661

fbshipit-source-id: 2b44ddeb35fd97a2f6d9b26b9ef23841602b11f5