Implement UntypedDefBehavior::SkipAndInferReturnAny
Summary:
This diff impements the skip-and-infer-return-any mode officially required
by PEP 484, in which functions that lack annotations are inferred to return
the gradual type `Any` and also have their function bodies skipped for
type checking.

One consequence of skipping function bodies is that we cannot do any type
analysis of attribute assignments, which means inference for `self.<attr>`
assignments in untyped functions has to be done in a gradually typed way;
here we implement the logic to infer the existence of all such attributes,
but with type `Any`. Noteably, because our analysis of annotations depends
on scope, we will ignore any type annotations on `self.<attr>` assignments
when in an unchecked function body.

Reviewed By: rchen152

Differential Revision: D74537912

fbshipit-source-id: 89ad873f7b7f9e63a010e7fd43102e43d3c31769