Add a stress test of more untyped defs
Summary:
One issue I ran into in my untyped defs implementation was that
the first version didn't correctly bing parameters (because these
live in the inner scope I want to skip, it was natural to skip
them and I didn't realize that they are also used by the solve
function for `Binding::FunctionDefinition`, so my initial
code would panic but only on functions with params).

I figured it's worth having a stress test of various combinations
of parameters, including ones that use the self type, to be
safe.

Reviewed By: yangdanny97

Differential Revision: D74537913

fbshipit-source-id: ad3381abb4cee84feb1e77de88e2b44e40377827