Add docs for `untyped_def_behavior`
Summary:
We can now emulate mypy's `check_untyped_defs`, which is widely used in
projects (including Pytorch) as well as the PEP 484 (and Pyre1) behavior of
skipping untyped defs. Add docs describing how this works, and detailing that
skipping checks will impact both attribute type inference and the availability
of language server features that depend on analyzing the code.

Reviewed By: yangdanny97

Differential Revision: D74550880

fbshipit-source-id: bb68d8fbb9c90e5e66a1c30477420c04b256b833