Implement `@no_type_check` support via SpecialExport
Summary:
We use SpecialExport when we need to resolve an identity at binding time, and
skipping function bodies is a binding-time decision.

The behavior when this decorator is present is the same as the behavior of
untyped defs when we have `untyped_def_behavior = "skip-and-infer-return-any"`.

Fixes https://github.com/facebook/pyrefly/issues/90

Reviewed By: yangdanny97

Differential Revision: D74582658

fbshipit-source-id: 9aad917d04e6d9121abbc7634ed60d32b8fabc2b