Config based `populate_all_project_files`
Summary:
In this diff, I implemented a config based `populate_all_project_files`, so that find-references can be enabled easily, or even enabled by default in the future.

Since during server init time, we don't necessarily know what are the projects we are working with, and we might always receive new projects in the future. Therefore, this is done during handling of `textDocument/didOpen`. I used `config.directory` to find the config, since this won't fallback to the default config. I used nothing from workspace, because it might be opened at the root of a big project.

Right now, I have two variants of indexing `LazyNonBlockingBackground` and `LazyBlocking`.  `LazyNonBlockingBackground` and `LazyBlocking` only differ in that `LazyNonBlockingBackground` will run the population in another thread. I believe `LazyBlocking` is only useful for tests, because we want determinism. Eventually, I want to make `LazyNonBlockingBackground` the default.

Reviewed By: kinto0

Differential Revision: D74591773

fbshipit-source-id: 849b82b5be3d80d0e781e63e5cb2ba2a253c6277