add disableTypeErrors config
Summary:
There's a chance users might want pyrefly for only language services, not type errors. In that case, they can use this setting to disable type errors. This should not be recommended as the settings only apply to one user: users should disable type errors via a project config if they don't want to see type errors.

In the future, we can switch to reusing `python.analysis.typeCheckingMode`. But currently I don't feel comfortable reusing it since our error levels might not align and it says pyright in the description.

Reviewed By: stroxler

Differential Revision: D74589812

fbshipit-source-id: 76e86209f1e5fd7b5902315bf3e399fe8bca5304