Fix spacing for landing page
Summary:
The spacing doesn't look great for the new landing page since there's a bunch of spacing around the Pyrefly header, and no spacing after the performance chart. This diff moves the spacing into `PyreflySection` so that we guarantee that extra spacing is added before the first section and after the last section.

There's also some code remenants in landingPageSection.tsx which are no longer relevant.

Reviewed By: stroxler, ndmitchell

Differential Revision: D74624640

fbshipit-source-id: 43a7544d17688085d716aec33727ceb120c48a8c