Add a helper type `MethodThatSetsAttr`
Summary:
Without a type, the (Name, bool) tuple is awkwardly floating in public
APIs, which Jia noted in diff review.

Giving it a named identity helps, and allows us to name the bool in the type.

This also let me tidy up doc comments and focus less on what the data is
and more on how it is used and how the ordering of method iteration is important
for Pyrefly behavior.

Reviewed By: grievejia

Differential Revision: D74337032

fbshipit-source-id: 8084dc68904ba476f7a74776b3c912d1577789bb