adjust error message and binding for AnnotationMismatch
Summary:
The only place this error code and binding are used is when comparing two definitions for the same variable.

Note that they don't have to be in the same branch of control flow; it will also work for something like this:

```
if y == 0:
  x: int = 1
else:
  x: str = ""
```

in which case the `x: str` is the "new" definition.

I wonder: does it make sense to have this on by default? what's the harm if we let people redefine variables?

closes https://github.com/facebook/pyrefly/issues/34

Reviewed By: Solumin

Differential Revision: D74662619

fbshipit-source-id: 916b1c6151b53d478bc95ad4cc5a6b4906abf4e9