add check_ignoring_loads_expectations
Summary: if we are to test contents that have changed without modifying the contents again, we need to be able to ignore the `# E: <>` contents we validate in `check`. The test I will add in the next diff must do this, but modifying the contents again will cause it to pass.

Reviewed By: SamChou19815

Differential Revision: D74669214

fbshipit-source-id: f37128e8022ff8a326e7f74de96ec06a5d730f68