Edit installation.mdx using inpage editor
Summary:
This diff has been automatically generated by the inpage editor.

Add docs for Pyrefly in CI

NOTE: If you want to update this diff, go via the preview link inside the static docs section below.
Ensure you are editing the same page that was used to create this diff.

Reviewed By: stroxler

Differential Revision: D74657180

fbshipit-source-id: e088b42359a7eddfb178f3baf7fa9ad1e1813747