normalize RHS of type aliases with untype_opt
Summary:
There are a few places in the type alias code that checks for `Type::Type`, but not all valid type expressions resolve to that type, for example `T | None`, since `None` is valid as both an expression and a type.

We can use `untype_opt`, which handles None, union, and a few special cases, as a way to normalize the type forms.

fixes https://github.com/facebook/pyrefly/issues/190

Reviewed By: migeed-z

Differential Revision: D74673360

fbshipit-source-id: 553f22ffb9b80788a6f7b763378a06b6c09311bc