Create WeakArcId
Summary:
In upcoming diffs, I'd like to keep track of the `ConfigFile`s that are loaded, especially for LSP. We'll need something like this so that we can make sure we're watching all configs, `search_path`s, and `site_package_path`s that are loaded, even if they're outside of the workspace.

This diff creates the new `WeakArcId` type, which is equivalent to a [`sync::Weak`](https://doc.rust-lang.org/std/sync/struct.Weak.html), retaining the original id of the `ArcId` for hashing purposes.

I'm doing this because it will enable us to have a collection of all loaded `ConfigFile`s for watching purposes without actually owning them or requiring that they stay in memory if nothing else is using it.

Reviewed By: rchen152, kinto0

Differential Revision: D74614307

fbshipit-source-id: afb7d43bdd30862f909dbf32c4f77daffdaee915