Add loaded_configs cache
Summary:
In upcoming diffs, I'd like to keep track of the `ConfigFile`s that are loaded, especially for LSP. We'll need something like this so that we can make sure we're watching all configs, `search_path`s, and `site_package_path`s that are loaded, even if they're outside of the workspace.

This diff adds a new cache of configs that have been loaded from the LSP config loader `configure` function. We use `WeakArcId`s to make sure we don't keep them in memory (and continue watching!) this functionality even after all other functionality doesn't require the config to be in memory anymore.

The next diffs add functionality to drop unused configs and watch files related to the given config.

Reviewed By: kinto0

Differential Revision: D74614308

fbshipit-source-id: 89003867cc4bfa446e223079ddb2ab0b3c75f20c