Add search_path and site_package_path to watcher in LSP
Summary:
In upcoming diffs, I'd like to keep track of the `ConfigFile`s that are loaded, especially for LSP. We'll need something like this so that we can make sure we're watching all configs, `search_path`s, and `site_package_path`s that are loaded, even if they're outside of the workspace.

This diff watches all loaded `ConfigFile`s if they have a path, their `search_path`s, and `site_package_path`s.

Reviewed By: kinto0

Differential Revision: D74616467

fbshipit-source-id: 26dde5757d31ff29fe20d96b42cf8b5e49bd9271