Make config finder override function return the ArcId to use
Summary:
In upcoming diffs, I'd like to keep track of the `ConfigFile`s that are loaded, especially for LSP. We'll need something like this so that we can make sure we're watching all configs, `search_path`s, and `site_package_path`s that are loaded, even if they're outside of the workspace.

The first step to doing this is to update config finders to return an already `ArcId`'d `ConfigFile`, which we'll be able to keep track of. I'm doing this for two reasons:
1. fewer `clone()`s that need to be done to the config to get this working
2. I'm making a `WeakArcId` type in the next diff which will allow us to hold onto this but not require that the `ConfigFile` stays in memory.

I think this should be a pretty reasonable change, it shouldn't make things too much more complicated to be doing this in the config finder `configure` functionality instead of having the resulting `ConfigFinder` object.

Reviewed By: kinto0, rchen152

Differential Revision: D74614305

fbshipit-source-id: c7a92d05fae73bd59dda0888aaa7333b9965022e