Add tooltip for type check speed as caveat
Summary:
Following stats review, the ask is to add caveat on the hardware we used to measure Pyrefly speed so that we have some way to explain if external users start complaining that they don't see the same performance. The list of caveats that we need to add is included in this task: T220437744. This diff adds the caveat for the type check over 1.85 million lines of code per second.

The rest of the caveats are included in the performance chart.

Reviewed By: stroxler

Differential Revision: D74696039

fbshipit-source-id: c5ff890b43cc27f6a2aa9c4c2859389b32334dc3