Check forced variables the right way around
Summary:
If we assume variable `x` is `int | bool`, but later realise it must be `bool`, that is OK - we used an over-conservative approximation. But if we later realise it must be `bool | str | int` then we have a problem. Flip the order of the subset test to ensure that.

Add the test from the  github issue.

Reviewed By: rchen152

Differential Revision: D74669118

fbshipit-source-id: 4f258a6b5d265755526f3ec1fda1ab65093dff44