Lazy load code snippet only when it comes into view port
Summary:
It's currently extremely slow to load the learn pages, where there are about 8 code snippets per page, taking about 2-3 seconds to load the page after clicking on it. This is a pretty bad UX.

Updating the code for codesnippet so that the code snippet will only be rendered when we are about to see it in the view port. This allows the page to load much faster

Reviewed By: stroxler

Differential Revision: D74700255

fbshipit-source-id: e1110af0be44cd19dae118e9abb5cefa40b399c9