update link to learn python typing docs
Summary: Reading over the content on the introduction page, we currently only talk about the link to "python typing for beginners" page. This is a bit confusing since we actually designed two pages, one for python beginners, and one for more experience python developers. Modifying the content here to include links to both pages.

Reviewed By: yangdanny97

Differential Revision: D74698216

fbshipit-source-id: e47d5a5d8c63471b62252e3ac62d735b8830805d