Add a placeholder README for our PyPI project
Summary:
Eventually we should figure out how to show our real README on pyrefly's PyPI page, but it's a little bit tricky because maturin errors on a "../README.md" path when building a source distribution.

For now, I've created a placeholder README by copying the disclaimer about pyrefly being under development (I used David's proposed new wording in D74402019) and the first paragraph of the intro from the website, just to get rid of the big "The author of this package has not provided a project description" message, which looks pretty bad.

Reviewed By: stroxler

Differential Revision: D74737366

fbshipit-source-id: 46acd64088b4ee5b83913aaebc1593662098f6ca