Add a simpler test case hinting at some match issues
Summary:
I'm investigating issue 197, where we panic in a specific match example.

I'm still working to understand what's happening, but even in this simpler
example there's something pretty fishy: we ought to have a clear error message
indicating that you can't do two name matches, becasue only the first one
is reachable.

We do have a preexisting example that hits the same issue, but it's a much
bigger example, having a tiny one to focus on is helpful. Also, I am starting
a `pattern_match` test module where I intend to eventually gather most `match`
related tests (they are scattered pretty widely right now) but I'm not going
to move all the tests just yet because that would lead to merge conflicts, at
the moment I just need an obvious place to dump *new* tests.

Instead, we seem to be attempting to analyze the second match as if it were
something other than a name match, leading to an internal error (not a panic,
but not good either!) because we try to look up the name `y`, when in fact
this is an (erroneous) attempt to *define* the name `y`.

I suspect the cause of the bad lookup may be related to a mismatch between
our static and flow environment bindings; we also ought to be producing a crystal
clear error message like the runtime does (which we are once this change is stacked
on top of the ruff semantic syntax checking change)

Reviewed By: yangdanny97

Differential Revision: D74810611

fbshipit-source-id: 83f7c26540df170f979f1f77f1f67478d5c1e978