Remove break on exhaustive match
Summary:
Fixes https://github.com/facebook/pyrefly/issues/197

It's true that an exhaustive match case makes any later cases
unreachable (and after enabling ruff semantic syntax checking, we
even get a syntax error that matches what the runtime would give).

But that doesn't mean our type checker should stop analyzing the
code - in general we've concluded that truncating control flow is
not a good user experience, because code in the middle of edits
will often enter states where some code is unreachable, and truncating
analysis like this leads to weird effects like all IDE features being
turned off in an unreachable branch (as far as Pyrefly is concerned,
unreachable code that we skip might as well be whitestpace).

In addition, it's asking for trouble: it's much more complicated
to break out of control flow, and `definitions.rs` doesn't do it.
As a result, in simple code like our test
```
match 42:
    case x:
        pass
    case y:
        pass
```
the *static* environment will actually have both `x` and `y`,
but the *flow* environment will break before we get to `y`.

This inconsistency means that we'll potentially panic if we ever try to look up
`y`. The only fixes are to either truncate analysis in `definition.rs`, or *not*
truncate analysis in `pattern.rs`. I'm opting for the latter given the reasoning
above.

Note that in addition to eliminating the internal error on the `case y` line
of our example, this diff allowed me to uncomment the `print(y)`; we've fixed
the panic.

We also expose an incorrect `assert_type` that was preexisting; this wasn't
caught because control flow analysis terminated but now we analyze the
`assert_type` so it has to be correct.

--------------

An open question is whether every unreachable match ought to include a type
error. I'm leaning toward no for now: we'll already get an error (matching the
Python syntax error we get when trying to run the code) on the case that makes
later cases unreachable. In my opinion that makes throwing a bunch of other
errors noisy and not necessary, especially since the error isn't *really*
suppressable (they could suppress it but the runtime is going to crash).

Reviewed By: yangdanny97

Differential Revision: D74811486

fbshipit-source-id: 1d695fc31111ff765c3038accda1e7d2d74dfcdd