bring back migrating-to-pyrefly page
Summary:
`migrating-to-pyrefly` was removed in D74746839 but this page is unfortunately linked in our blog, and we are now hitting a 404 when accessing this page (https://pyrefly.org/en/docs/migrating-to-pyrefly/). Bringing this page back so that we don't get a 404 from the blog.

Making this a separate page so that we are able to keep the sidebar categories as non-clickable.

Reviewed By: ndmitchell

Differential Revision: D74814375

fbshipit-source-id: f93171d912f65cbcda4b1ed2a3e89022d4b383fc