Fix incremental panic
Summary:
This should fix https://github.com/facebook/pyrefly/issues/199

Unfortunately the issue is not easily reproducible, but I'm fairly certain
I understand what is happening (we're mixing type values coming from
incompatible versions of the same module, and getting arity mismatches)

The only reported panic caused by this happens in `subset`, so we can just
make this robust. The only obvious way to deal with it is to say that if arities
mismatch, then the two classes are from different versions of the code and
for now they are not linearly orderable (trying to do anything else is tricky
because we'd have work out downstream logic).

Reviewed By: ndmitchell

Differential Revision: D74819844

fbshipit-source-id: 718dea117b61001ca02eb7934af433a452cb81e7