set color for performance comparison button and pip install copy button
Summary: Neil reported seeing the blue color for performance comparison button and the pip install copy button. This seem to be because they are using the default colors on phone, and we didn't define the colors for these elements. Setting these to color-text to be consistent with the rest of the page.

Reviewed By: MaggieMoss

Differential Revision: D74818856

fbshipit-source-id: 8213b48705d0a9d3c03adb9549445d1237656433