Move logic to get globs for config file to util function
Summary:
In upcoming diffs, I'd like to keep track of the `ConfigFile`s that are loaded, especially for LSP. We'll need something like this so that we can make sure we're watching all configs, `search_path`s, and `site_package_path`s that are loaded, even if they're outside of the workspace.

This diff pulls out the functionality to get globs for a given `ConfigFile` we'll use for watching into a util function we can reuse when dealing with newly opened or changed config files.

Reviewed By: kinto0

Differential Revision: D74618565

fbshipit-source-id: 00e6cfa1c2e58e4af71187a25b5f42e29a0328e2