Add a hover test illustrating a bug unwrapping
Summary:
When we unwrap types (aside: I think in a Rust codebase this might not
be a great name, we're really decomposing... might be worth a rename),
we do it by:
- creating fresh vars for the generic params of the type we want to
  decompose
- doing a subset check
- if it passes, espanding the vars and returning

This usually works great: it will handle subtyping and such, so
on and so forth.

The problem is that in a gradual type system, the subset check
can pass without actually pinning the var.

This new test illustrates the problem: if we have an `Any` and we
try to unwrap it as an `Awaitable` (which we have to do implicitly
for `await` because we need to get the type argument), we never
force the vars.

Fix coming in the next diff, I wanted to make it visible in the tests
first. Note that we probably want a way to reveal types without
forcing them - I'm going to look at reveal_type later this week, I
think it should just show the var.

Reviewed By: ndmitchell

Differential Revision: D74971503

fbshipit-source-id: 34de6969a7259bcae87fcc6e54b62405bf7f1538