Start forcing, not expanding, vars in `unwrap.rs`
Summary:
See the previous diff for a bigger discussion - basically, if we don't
force vars in this scenario, comparisons to Any can cause us to pass
a subset_eq check while failing to pin the vars.

The solution is to force, so any upinned vars become `Any` (which is
correct - they are coming from some kind of gradual type, so this is
what we want!)

This diff is the fix.

Reviewed By: ndmitchell

Differential Revision: D74971502

fbshipit-source-id: 9dfaa01467ee54b8caa203490c31760180f262bc