Add a proper `return_any` test of await
Summary:
I realized our original test isn't actually doing quite what I thought it
was - we were awaiting a `Coroutine[Any, Any, Any]`, not an `Any` directly.

I'm actually slighty fuzzy on what happens in the case we were looking at, but
we know that we fixed it. This diff adds the case of *directly* awaiting `Any`
which is what I thought I was doing before. The good news is, it works as
expected :)

Reviewed By: ndmitchell

Differential Revision: D75012112

fbshipit-source-id: 7e884d697ac071999095e6d61d11222801d613f7