Make `reveal_type` expand, not deep force
Summary:
It is not 100% clear what the end behavior of reveal_type should be.

For power users and for Pyrefly team members, it would be good to reveal
what the type is at the moment we call it, which is what this diff does:
we should *expand* rather than force so that vars that aren't pinned yet
are visible as vars.

For average end users, we possibly should eventually get behavior closer
to hover where vars that *never* get pinned are visible as vars and vars
that eventually get pinned show the pinned data.

But deep_force in a Binding cannot achieve this - deep forcing eagerly
doesn't actually do what I described above, it does something much worse:
it takes any existing vars and *side-effect*s the environment, forcing
them to gradual types.

So if we want to produce the behavior described above, I suspect the way
to do it would be to make a custom binding that is resolved *last* (after
the rest of the module was type checked). This may be a good idea, but is
also quite a bit of work and I'm not even certain yet whether it's better for
users so for now I will just use `expand`. And it might even be easier to do
in a future state of Pyrefly, which is more reason to defer for now.

Reviewed By: connernilsen

Differential Revision: D75006615

fbshipit-source-id: 3cfb8fff48e44be4653735b98fc67400e8f219b5