document symbols fails with unnamed symbols
Summary:
We send back a  name = "" symbol sometimes when users are typing and we have a parse tree that is not complete. Example:
```
def <cursor>
```
This causes the language client to crash: it doesn't like document symbols with a falsy name.

This diff fixes up any "" name and calls it "unknown". There will be tests and a better fix soon.

note: it was hard to repro this issue on the internal extension because it errored in the terminal but not in UI. I will see if I can fix that to catch future regressions better.

Reviewed By: ndmitchell

Differential Revision: D75067540

fbshipit-source-id: aeeeec15abe3fb09b542e0ddc09866d236b65784