Rename property to facet in narrowing
Summary:
Properties (as in Python properties, which is a name we cannot change)
are relevant to narrowing, which makes "property narrowing" a tricky term
to use for narrowing attributes + indexes.

After some discussion, we decided "facet" is a good name because it isn't
overloaded with anything else (I floated "access", but that sounds too
much like a verb, and "element" as alternatives).

This diff does the rename: a narrowed subtree thing lives at a "facet chain",
an entry in the chain is a "facet", and a slice of entries in the chain are
"facets".

Reviewed By: grievejia

Differential Revision: D75071371

fbshipit-source-id: e545f30c27049c1385763ba7af88400f66338beb