Add a simple emacs config
Summary:
Emacs is a somewhat common editor, we've had a couple folks ask about using it.

There are many ways to configure emacs, but the simplest approach is to use `eglot`
which is the most standard of the language server protocol implementations and actually
ships with recent versions of emacs.

I provide both the simplest imperative initialization, and what it looks like inside
a `use-package` block (`use-package` is the most common package manager).

Reviewed By: rchen152

Differential Revision: D75083501

fbshipit-source-id: 516dae18f350df048ac6c9e1a6835028105a5ca0