Add Info, Warn as error severities
Summary:
We pick the severity based on the kind, which I think makes sense - we should always have a good kind, and having one kind with multiple severities seems bad. Also lets us have a config to change severity of each warning, not just to disabled.

For now, just mark RevealType as Info.

Reviewed By: grievejia

Differential Revision: D75090898

fbshipit-source-id: 8a7d0a29026ee5dc3129d92bb9139c975b1bfd41