Narrow bindings should preserve FlowStyle
Summary:
We were overriding the `FlowStyle` of the original binding when creating the narrowing bindings. This is undesirable because we rely on `FlowStyle` downstream to determine whether the original binding is an import / initialized / etc., and losing that information cause Pyrefly to misbehave.

This patch should fix https://github.com/facebook/pyrefly/issues/274

Reviewed By: rchen152

Differential Revision: D75081356

fbshipit-source-id: 983274f8926b319f0f77717c89c2e0bd7c8ca2f9