Move check for when to ignore a class's empty tparams into ClassMetadata
Summary:
This will make it easier to extend this check to handle another case in which we shouldn't report bad-specialization errors.

We have to put the has_unknown_tparams flag in ClassMetadata (rather than Class or TParams) because it uses has_base_any, which isn't computed until we get the types of the base classes for ClassMetadata.

Reviewed By: ndmitchell

Differential Revision: D75105959

fbshipit-source-id: 30d45500504d062befa9f87d823ab1814b6ee075