Fix `TypeInfo::join` when at least one entry has no facets
Summary:
For efficiency, `TypeInfo` represents empty facets as `None` rather
than carrying an empty map around. This is probably worth it because
every binding makes a `TypeInfo`, so we get both a smaller data structure
with fewer allocs and much cheaper operations in the normal case of no
narrows.

But it does make the `join` logic quite complex, and here I missed a case:
we correctly drop the facets if there are no overlapping names in a bunch
of `NarrowedFacets` branches, but we forgot to make sure that none of the
`TypeInfo` branches has `facets: None`.

The best way to check for that given the exiting code structure is to check
the length of the flattened `facets` branches; if any of them were None this
will be smaller than the length of the `TypeInfo` branches and we should
produce `None`.

Reviewed By: ndmitchell

Differential Revision: D75140234

fbshipit-source-id: c4839d0cf69df29cce883075763c4022b46bb93e