Add a test case of a type_info bug
Summary:
When we join a TypeInfo with no facet narrows to one with one or
more facet narrows, we are incorrectly preserving the narrows.

That causes this example to incorrectly preserve narrows at the end of the
`test` function.

The root cause is a missing count check after a `filter_map`, I'll
stack the fix on top.

Reviewed By: ndmitchell

Differential Revision: D75140235

fbshipit-source-id: 5d4c9548462b0fc006c227759f1735bde0d82447