Do not redefine `function_idx`
Summary:
This line has confused me for a while but I didn't have time to work through
it. Now that I have, I've realised it's harmless at runtime but misleading: this
index is going to be identical to the `function_idx` we computed at the top (it
had better be, otherwise things would break badly because we passed it to
`self.function_body` and a mismatch will cause panics later!)

The intent of the code is clearer if we just ignore the return value of
`self.table.insert`.

Reviewed By: ndmitchell

Differential Revision: D75155236

fbshipit-source-id: b257bed2b290f112ba3a4c6bc05855bff48296d8