Distinguish between pre and post inference variance
Summary:
Right now, variance is optional.

I think we should change that by creating a pre-inference variance. If we have PEP695 syntax, the pre-inference can be undefined. When we infer the type, we change the representation to the post variance representation.

I think this representation is better so we don't have to deal with all those optional.

Next step: ~~I am hoping to thread a map through the typechecker, which will carry a variance environment with the inferred variants.  Instead of converting variance from pre to posst on the fly, we can read from the map.~~

I don't think we need to literally thread a map. In pyre1, we used a call-back. I will look through the code to determine exactly how this should be structured but the high level idea is to be able to get variance by going through the variance inference algorithm instead of inferring it on the fly like we're doing now.

Reviewed By: yangdanny97

Differential Revision: D75091014

fbshipit-source-id: f9d61236d0d396417d5931b8948a8422daf661eb