Check that condition.__bool__ is callable in if and while statements
Summary:
In `if <condition>` and `while <condition>`, the value of the condition
expression is coerced to bool by calling its `__bool__` method, so we need to
check that it is callable.

Follows similar checks for expression-if and explicit calls to `bool()`, and
modifies the error message to show the type of the invalid `__bool__`
attribute.

Reviewed By: stroxler

Differential Revision: D75118659

fbshipit-source-id: 279b9b2b02e1a529d7eb2458a3125bb2fed37a5b