Rename scope inners + vars
Summary:
This is a very surface-level change, but it actually addresses a longstanding
pain point I've had, which is that calling a class *scope* the "class body"
kept getting me confused, especially in the more complex bindings functions
where there are sometimes a large number of variables to keep track of (at
which point good names matter more).

I also never much liked the names `.*Inner`, becasue this naming convention is
normally associated with an inner sub-struct of a datatype (e.g. `ClassInner`
is a substruct of `Class`), whereas these represent the contents of one particular
enum variant, and we don't usually call those `.*Inner`.

In addition, I don't think `ClassBody` was ever a great name anyway for a couple
reasons:
- the notion of "class body" is already well-defined to be something that isn't
  exactly a scope (it is a chunk of an AST, and we have a lot of helpers named
  this way already)
- in addition, there's no notion of "class scope" that is anything *other* than
  a class body. So there's no reason to use a long name that could be a little
  confusing, "class scope" is already unambigous.

This diff updates both of the datatype names, and audits variable names in various
bindings functions to be consistent as well.

Reviewed By: yangdanny97

Differential Revision: D75164686

fbshipit-source-id: c3db7baa13189d286339ff893bb0bea1f5501852